Changes:
Megan and I deleted the personList class and added a new DataStorage
class that loads and saves person data, as well as addes a new person
through the gui. 

We also made a csv that hold the last used Id and Connection numbers and
created methods in DataStorage that load and save current and new
numbers.