Changes:
Made it so the Add Person GUI is now Add/Edit and takes in a parameter
of a person object. If null then will open the blank add gui, else will
pull up the add gui with the selected person it it. 

So editing and saving of a person works completely!
As far as editing and saving of a connection...not so much.